refactor: extract business logic into services modules (#17)
Create services/ layer between command handlers and core modules.
Command handlers are now thin IPC adapters. Socket.rs session create
now calls services instead of duplicating logic.

Closes #17
