+++ src-tauri/src/commands/projects.rs
@@ -1,4 +1,5 @@
 use crate::projects::Project;
+use crate::services::projects as svc;
 use crate::state::AppState;
 
 #[tauri::command]
@@ -24,37 +25,16 @@ pub(crate) fn rename_project(id: String, name: String, state: tauri::State<AppSt
 }
 
 #[tauri::command]
-pub(crate) async fn set_session_project(
-    session_id: String,
-    project_id: Option<String>,
-    state: tauri::State<'_, AppState>,
-) -> Result<(), String> {
-    let handle = state.session_handle.clone();
-    handle.set_project(&session_id, project_id).await.map_err(|e| e.to_string())?;
-    Ok(())
+pub(crate) async fn set_session_project(session_id: String, project_id: Option<String>, state: tauri::State<'_, AppState>) -> Result<(), String> {
+    state.session_handle.set_project(&session_id, project_id).await.map_err(|e| e.to_string())
 }
 
 #[tauri::command]
 pub(crate) fn get_project_notes(project_id: String) -> Result<String, String> {
-    let path = notes_path(&project_id);
-    if path.exists() {
-        std::fs::read_to_string(&path).map_err(|e| format!("Failed to read notes: {}", e))
-    } else {
-        Ok(String::new())
-    }
+    svc::get_notes(&project_id).map_err(|e| e.to_string())
 }
 
 #[tauri::command]
 pub(crate) fn set_project_notes(project_id: String, content: String) -> Result<(), String> {
-    let path = notes_path(&project_id);
-    if let Some(parent) = path.parent() {
-        std::fs::create_dir_all(parent)
-            .map_err(|e| format!("Failed to create notes dir: {}", e))?;
-    }
-    std::fs::write(&path, &content).map_err(|e| format!("Failed to write notes: {}", e))
-}
-
-fn notes_path(project_id: &str) -> std::path::PathBuf {
-    let base = dirs::config_dir().unwrap_or_else(|| std::path::PathBuf::from("."));
-    base.join("roux").join("notes").join(format!("{}.txt", project_id))
+    svc::set_notes(&project_id, &content).map_err(|e| e.to_string())
 }
